fix: Breaking changes from content file structure changes
